# docs/openapi.json
{
  "components": {
    "schemas": {
      "ActionItemAssignRequest": {
        "additionalProperties": false,
        "properties": {
          "due_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "owner_id": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "reason": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "ActionItemAssignRequest",
        "type": "object"
      },
      "ActionItemCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "dataset_type": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Type"
          },
          "domain": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Domain",
            "type": "string"
          },
          "kind": {
            "maxLength": 60,
            "minLength": 1,
            "title": "Kind",
            "type": "string"
          },
          "reference_key": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Key"
          },
          "scope_code": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Code"
          },
          "title": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "domain",
          "kind",
          "title"
        ],
        "title": "ActionItemCreateRequest",
        "type": "object"
      },
      "ActionItemListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ActionItemResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ActionItemListResponse",
        "type": "object"
      },
      "ActionItemReopenRequest": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "ActionItemReopenRequest",
        "type": "object"
      },
      "ActionItemResolveRequest": {
        "additionalProperties": false,
        "properties": {
          "comment": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Comment",
            "type": "string"
          }
        },
        "required": [
          "comment"
        ],
        "title": "ActionItemResolveRequest",
        "type": "object"
      },
      "ActionItemResponse": {
        "additionalProperties": false,
        "properties": {
          "assigned_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned At"
          },
          "assigned_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned By"
          },
          "assignment_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignment Reason"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "dataset_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Type"
          },
          "domain": {
            "title": "Domain",
            "type": "string"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_overdue": {
            "title": "Is Overdue",
            "type": "boolean"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "reference_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Key"
          },
          "resolution_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Comment"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "resolved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved By"
          },
          "scope_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Code"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "domain",
          "kind",
          "title",
          "dataset_type",
          "scope_code",
          "reference_key",
          "status",
          "owner_id",
          "due_date",
          "created_by",
          "created_at",
          "assigned_by",
          "assigned_at",
          "assignment_reason",
          "resolved_by",
          "resolved_at",
          "resolution_comment",
          "is_overdue"
        ],
        "title": "ActionItemResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PortfolioItem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "latest_published_report_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Published Report Date"
          },
          "latest_published_snapshot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Published Snapshot Id"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "reporting_currency": {
            "title": "Reporting Currency",
            "type": "string"
          }
        },
        "required": [
          "code",
          "name",
          "reporting_currency",
          "latest_published_report_date",
          "latest_published_snapshot_id"
        ],
        "title": "PortfolioItem",
        "type": "object"
      },
      "PortfolioListResponse": {
        "additionalProperties": false,
        "properties": {
          "combined_report_dates": {
            "items": {
              "type": "string"
            },
            "title": "Combined Report Dates",
            "type": "array"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PortfolioItem"
            },
            "title": "Items",
            "type": "array"
          },
          "report_date_mismatch": {
            "title": "Report Date Mismatch",
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "combined_report_dates",
          "report_date_mismatch"
        ],
        "title": "PortfolioListResponse",
        "type": "object"
      },
...
      "SnapshotListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SnapshotSummary"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "SnapshotListResponse",
        "type": "object"
      },
...
      "SnapshotSummary": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "import_id": {
            "title": "Import Id",
            "type": "string"
          },
          "portfolio": {
            "title": "Portfolio",
            "type": "string"
          },
          "report_date": {
            "title": "Report Date",
            "type": "string"
          },
          "source_upload_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Upload Id"
          },
          "status": {
            "enum": [
              "draft",
              "validating",
              "validated",
              "approved",
              "published",
              "failed",
              "rejected",
              "superseded",
              "withdrawn"
            ],
            "title": "Status",
            "type": "string"
          },
          "value_label": {
            "title": "Value Label",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "import_id",
          "portfolio",
          "report_date",
          "version",
          "status",
          "value_label"
        ],
        "title": "SnapshotSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OidcBearer": {
        "description": "Production OIDC access token. Development headers are local-only.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Portfolio Operations Insight API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/action-items": {
      "get": {
        "operationId": "listActionItems",
        "parameters": [
          {
            "in": "query",
            "name": "domain",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Domain"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionItemListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "List Items",
        "tags": [
          "action-items"
        ]
      },
      "post": {
        "operationId": "createActionItem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionItemCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Create Item",
        "tags": [
          "action-items"
        ]
      }
    },
    "/api/v1/action-items/{item_id}": {
      "get": {
        "operationId": "getActionItem",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Get Item",
        "tags": [
          "action-items"
        ]
      }
    },
    "/api/v1/action-items/{item_id}/assign": {
      "post": {
        "operationId": "assignActionItem",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionItemAssignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Assign Item",
        "tags": [
          "action-items"
        ]
      }
    },
    "/api/v1/action-items/{item_id}/reopen": {
      "post": {
        "operationId": "reopenActionItem",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionItemReopenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Reopen Item",
        "tags": [
          "action-items"
        ]
      }
    },
    "/api/v1/action-items/{item_id}/resolve": {
      "post": {
        "operationId": "resolveActionItem",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionItemResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Resolve Item",
        "tags": [
          "action-items"
        ]
      }
    },
...
    "/api/v1/portfolios": {
      "get": {
        "operationId": "listPortfolios",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "List governed portfolios and publication dates",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/portfolios/{code}/snapshots": {
      "get": {
        "operationId": "listPortfolioSnapshots",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_unpublished",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Unpublished",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_superseded",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Superseded",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "List portfolio snapshots",
        "tags": [
          "catalog"
        ]
      }
    },
...
  }
}